Extract Element definition to new file.
